+++ CHANGELOG.md
@@ -144,6 +144,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 - get query results' snippets from corpus + format (#178)
 - fix ranking tokenization (#182)
 - add circuitbreakers (#185)
+- allow for missing mithril docs (#184)
 
 ## [0.1.0] - 2025-02-07